Add initial images Change columns in `/Home/Index`

+++ BulkyWeb/Areas/Admin/Controllers/ProductController.cs
@@ -156,7 +156,10 @@ public class ProductController : Controller
 
                 if (System.IO.File.Exists(oldImagePath))
                 {
-                    System.IO.File.Delete(oldImagePath);
+                    if (!oldImagePath.Contains("initial"))
+                    {
+                        System.IO.File.Delete(oldImagePath);
+                    }
                 }
 
                 // Delete imageDirectory if no files are remaining.
